corrected bug in refactored fragmented message receiving
git-svn-id: https://iiop-net.svn.sourceforge.net/svnroot/iiop-net/trunk@31 b113d865-ab28-0410-9ae9-d2690f63996d

+++ IIOPNet/IIOPChannel/GIOPHeader.cs
@@ -83,7 +83,7 @@ namespace Ch.Elca.Iiop {
         /// <summary>the length of the header in bytes</summary>
         internal const int HEADER_LENGTH = 12;
     	
-    	internal const int FRAGMENT_MASK = 0x02;
+    	internal const byte FRAGMENT_MASK = 0x02;
         
         #endregion Constants
         #region IFields
